put dom di in its own package
