+++ packages/component/main.ts
@@ -1,9 +1,8 @@
-export { defineEnvironment, clearEnvironment, getEnvironmentRef } from './lib/environment';
 export { State } from './lib/state';
 export { handle } from './lib/handle';
 export { property, PropValidator, PropValidationError } from './lib/property';
 export { component, ComponentDef, RenderCtx, RenderDef, getComponentDef } from './lib/component';
-export { JoistElement, InjectorBase, get, withInjector, withPropChanges } from './lib/element';
+export { JoistElement, withPropChanges } from './lib/element';
 export {
   OnAttributeChanged,
   OnConnected,
@@ -14,3 +13,5 @@ export {
   OnComplete,
   HandlerCtx,
 } from './lib/lifecycle';
+
+export * from '@joist/di/dom';
