+++ packages/di/.gitignore
@@ -1,1 +1,2 @@
 target
+!typings.d.ts
