+++ packages/di/typings.d.ts
@@ -1,0 +1,16 @@
+declare interface CSSStyleSheet {
+  replace(text: string): Promise<CSSStyleSheet>;
+  replaceSync(text: string): void;
+}
+
+declare interface Document {
+  adoptedStyleSheets: readonly CSSStyleSheet[];
+}
+
+declare interface ShadowRoot {
+  adoptedStyleSheets: readonly CSSStyleSheet[];
+}
+
+declare interface HTMLElement {
+  connectedCallback(): void;
+}
